+++ core/src/commonMain/kotlin/at/asitplus/awesn1/Asn1Elements.kt
@@ -18,7 +18,7 @@ import kotlin.native.ObjCName
  * Base ASN.1 data class. Can either be a primitive (holding a value), or a structure (holding other ASN.1 elements)
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 sealed class Asn1Element(
     val tag: Tag
 ) {
@@ -479,7 +479,7 @@ val Asn1Null = Asn1Primitive(Asn1Element.Tag.NULL, byteArrayOf())
  * ASN.1 structure. Contains no data itself, but holds zero or more [children]
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 sealed class Asn1Structure(
     /**
      * The tag identifying this structure
@@ -688,7 +688,7 @@ sealed class Asn1Structure(
  * Explicit ASN.1 Tag. Can contain any number of [children]
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1ExplicitlyTagged
 /**
  * @param tag the ASN.1 Tag to be used will be properly encoded to have [BERTags.CONSTRUCTED] and
@@ -742,7 +742,7 @@ internal constructor(tag: ULong, children: List<Asn1Element>) :
  * @param children the elements to put into this sequence
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1Sequence internal constructor(children: List<Asn1Element>) :
     Asn1Structure(Tag.SEQUENCE, children, sortChildren = false, shouldBeSorted = false) {
 
@@ -761,7 +761,7 @@ fun createAsn1Sequence(children: List<Asn1Element>): Asn1Sequence = Asn1Sequence
  * ASN1 structure (i.e. containing child nodes) with custom tag
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1CustomStructure private constructor(
     tag: Tag, children: List<Asn1Element>, sortChildren: Boolean, shouldBeSorted: Boolean
 ) :
@@ -849,7 +849,7 @@ class Asn1CustomStructure private constructor(
  * @param children the elements to put into this sequence
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
     Asn1Structure(Tag.OCTET_STRING, children, sortChildren = false, shouldBeSorted = false),
     Asn1OctetString {
@@ -878,7 +878,7 @@ class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
  * Cast to [Asn1OctetString] instead.
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1PrimitiveOctetString(content: ByteArray) : Asn1Primitive(Tag.OCTET_STRING, content),
     Asn1OctetString {
 
@@ -897,7 +897,7 @@ class Asn1PrimitiveOctetString(content: ByteArray) : Asn1Primitive(Tag.OCTET_STR
  * ASN.1 SET 0x31 ([BERTags.SET] OR [BERTags.CONSTRUCTED])
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 open class Asn1Set private constructor(children: List<Asn1Element>, dontSort: Boolean) :
     Asn1Structure(Tag.SET, children, !dontSort, shouldBeSorted = true) {
 
@@ -931,7 +931,7 @@ fun createAsn1Set(children: List<Asn1Element>): Asn1Set = Asn1Set(children)
  * @throws Asn1Exception if children are using different tags
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: List<Asn1Element>) :
     Asn1Set(children.also { it ->
         if (it.any { elem -> elem.tag != it.first().tag }) throw Asn1Exception("SET OF must only contain elements of the same tag")
@@ -941,7 +941,7 @@ class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: Lis
  * ASN.1 primitive. Holds no children, but [content] under [tag]
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 open class Asn1Primitive(
     tag: Tag,
     /**
@@ -1021,7 +1021,7 @@ open class Asn1Primitive(
  * (Your arbitrary bytes might inadvertently be valid ASN.1!)
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
-@Serializable(with = Asn1ElementStringSerializer::class)
+@Serializable(with = Asn1ElementFallbackBase64Serializer::class)
 sealed interface Asn1OctetString {
 
     /**